Refactor the epm/eps construction methods to reduce code duplciation
